order history providor refactored
