+++ src/providers/user/orderHistoryProvider/provider.tsx
@@ -1,37 +1,25 @@
-import { useCallback, useReducer } from "react";
+import { useCallback, useEffect, useReducer, useState } from "react";
 
 import * as queries from "../../../graphql/queries";
 import { useProfile } from "../profile";
+import { useSessionProvider } from "../sessionProvider/useSessionProvider";
 
 import { Provider } from "./context";
 import { ordersHistoryReducer, initialOrdersHistoryState } from "./reducer";
 import * as A from "./actions";
-import { SetOrder } from "./types";
+import { SetOrder, orderHistoryQueryResult } from "./types";
 
 import { UserAccount } from "@polkadex/orderbook/providers/user/profile/types";
 import { OrderCommon } from "@polkadex/orderbook/modules/types";
 import { fetchAllFromAppSync } from "@polkadex/orderbook/helpers/appsync";
 import { Utils } from "@polkadex/web-helpers";
+import { sortOrdersDescendingTime } from "@polkadex/orderbook/helpers/sortOrderDescendingTime";
+import { Ifilters } from "@polkadex/orderbook-ui/organisms";
 
 export const OrderHistoryProvider = ({ children }) => {
   const [state, dispatch] = useReducer(ordersHistoryReducer, initialOrdersHistoryState);
   const profileState = useProfile();
 
-  type orderHistoryQueryResult = {
-    u: string;
-    cid: string;
-    id: string;
-    t: string;
-    m: string;
-    s: string;
-    ot: string;
-    st: string;
-    p: string;
-    q: string;
-    afp: string;
-    fq: string;
-    fee: string;
-  };
   const account: UserAccount = profileState.selectedAccount;
 
   const fetchOpenOrders = useCallback(async (proxy_acc: string): Promise<OrderCommon[]> => {
@@ -60,6 +48,18 @@ export const OrderHistoryProvider = ({ children }) => {
     }));
     return orders;
   }, []);
+
+  const onOpenOrdersHistoryFetch = useCallback(async () => {
+    try {
+      if (account.tradeAddress) {
+        const transactions: OrderCommon[] = await fetchOpenOrders(account.tradeAddress);
+        dispatch(A.userOpenOrderHistoryData({ list: transactions }));
+      }
+    } catch (error) {
+      console.error(error);
+      dispatch(A.userOpenOrdersHistoryError(error));
+    }
+  }, [account.tradeAddress, fetchOpenOrders]);
 
   const fetchOrders = useCallback(
     async (proxy_acc: string, dateFrom: Date, dateTo: Date): Promise<OrderCommon[]> => {
@@ -97,19 +97,6 @@ export const OrderHistoryProvider = ({ children }) => {
     []
   );
 
-  const onOpenOrdersHistoryFetch = useCallback(async () => {
-    try {
-      if (account.tradeAddress) {
-        const transactions: OrderCommon[] = await fetchOpenOrders(account.tradeAddress);
-
-        dispatch(A.userOpenOrderHistoryData({ list: transactions }));
-      }
-    } catch (error) {
-      console.error(error);
-      dispatch(A.userOpenOrdersHistoryError(error));
-    }
-  }, [account.tradeAddress, fetchOpenOrders]);
-
   const onOrdersHistoryFetch = useCallback(
     async ({ dateFrom, dateTo, tradeAddress }) => {
       try {
@@ -124,17 +111,6 @@ export const OrderHistoryProvider = ({ children }) => {
     },
     [fetchOrders]
   );
-
-  const OnOrderUpdates = (setOrder: SetOrder) => {
-    try {
-      const order = processOrderData(setOrder);
-      dispatch(A.orderUpdateEventData(order));
-    } catch (error) {
-      console.log(error, "Something has gone wrong (order updates channel)...", error);
-      dispatch(A.orderUpdateEventError(error));
-    }
-  };
-
   function processOrderData(eventData: SetOrder): OrderCommon {
     const base = eventData.pair.base_asset;
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -157,6 +133,72 @@ export const OrderHistoryProvider = ({ children }) => {
       fee: eventData.fee.toString(),
     };
   }
+  const onOrderUpdates = (payload: A.OrderUpdateEvent["payload"]) => {
+    try {
+      const order = processOrderData(payload);
+      dispatch(A.orderUpdateEventData(order));
+    } catch (error) {
+      console.log(error, "Something has gone wrong (order updates channel)...", error);
+      dispatch(A.orderUpdateEventError(error));
+    }
+  };
+
+  const { dateTo, dateFrom } = useSessionProvider();
+  const usingAccount = profileState.selectedAccount;
+
+  const orderList = state.list;
+  const openOrders = state.openOrders;
+  const list = sortOrdersDescendingTime(orderList);
+  const openOrdersSorted = sortOrdersDescendingTime(openOrders);
+  const userLoggedIn = profileState.selectedAccount.tradeAddress !== "";
+
+  const [updatedList, setUpdatedList] = useState(list);
+  const [updatedOpenOrdersSorted, setUpdatedOpenOrdersSorted] = useState(openOrdersSorted);
+
+  useEffect(() => {
+    onOpenOrdersHistoryFetch();
+    onOrdersHistoryFetch({ dateFrom, dateTo, tradeAddress: usingAccount.tradeAddress });
+  }, [
+    usingAccount.tradeAddress,
+    dateFrom,
+    dateTo,
+    onOpenOrdersHistoryFetch,
+    onOrdersHistoryFetch,
+  ]);
+
+  const filterOrders = useCallback(
+    (filters: Ifilters) => {
+      if (filters?.onlyBuy && filters?.onlySell) {
+        setUpdatedList(list);
+        setUpdatedOpenOrdersSorted(openOrdersSorted);
+      } else if (filters?.onlyBuy) {
+        setUpdatedList(list.filter((data) => data.side?.toUpperCase() === "BID"));
+        setUpdatedOpenOrdersSorted(
+          openOrdersSorted.filter((data) => data.side?.toUpperCase() === "BID")
+        );
+      } else if (filters?.onlySell) {
+        setUpdatedList(list.filter((data) => data.side.toUpperCase() === "ASK"));
+        setUpdatedOpenOrdersSorted(
+          openOrdersSorted.filter((data) => data.side?.toUpperCase() === "ASK")
+        );
+      } else if (filters?.hiddenPairs) {
+        setUpdatedList(
+          list.filter((data) => {
+            return data.side.toUpperCase() !== "ASK" || data.side.toUpperCase() !== "BID";
+          })
+        );
+        setUpdatedOpenOrdersSorted(
+          openOrdersSorted.filter((data) => {
+            return data.side.toUpperCase() !== "ASK" || data.side.toUpperCase() !== "BID";
+          })
+        );
+      } else {
+        setUpdatedList(list);
+        setUpdatedOpenOrdersSorted(openOrdersSorted);
+      }
+    },
+    [list, openOrdersSorted]
+  );
 
   return (
     <Provider
@@ -164,6 +206,11 @@ export const OrderHistoryProvider = ({ children }) => {
         ...state,
         onOpenOrdersHistoryFetch,
         onOrdersHistoryFetch,
+        onOrderUpdates,
+        orders: updatedList,
+        openOrders: updatedOpenOrdersSorted,
+        userLoggedIn,
+        filterOrders,
       }}>
       {children}
     </Provider>
