+++ src/ui/organisms/Transactions/index.tsx
@@ -1,5 +1,5 @@
 import { DateRangePicker, defaultStaticRanges } from "react-date-range";
-import { useCallback, useMemo, useState } from "react";
+import { useCallback, useMemo, useState, useEffect } from "react";
 
 import * as S from "./styles";
 
@@ -46,14 +46,19 @@ export const Transactions = () => {
   const [trigger, setTrigger] = useState(false);
   const [isVisible, setIsVisible] = useState(true);
 
-  const orderHistory = useOrderHistoryProvider(filters);
-
+  const orderHistory = useOrderHistoryProvider();
+  const { filterOrders } = orderHistory;
   const userSession = useSessionProvider();
   const { dispatchUserSessionData } = userSession;
 
   // Filters Actions
-  const handleChangeHidden = (type: "hiddenPairs" | "onlyBuy" | "onlySell") =>
+  const handleChangeHidden = (type: "hiddenPairs" | "onlyBuy" | "onlySell") => {
     setFilters({ ...filters, [type]: !filters[type] });
+  };
+
+  useEffect(() => {
+    filterOrders(filters);
+  }, [filterOrders, filters]);
 
   const handleSelect = useCallback(
     ({ selection: { startDate, endDate } }) => {
